+++ e2e/tests/server-steps.js
@@ -341,13 +341,21 @@ step("Log has entry with key <key> and value <value>", async (key, value) => {
   assert(matches.length, `Could not find any log entry with key ${key} and value '${value}' but found ${values.map(v => `'${v}'`).join(', ')}`)
 })
 
-step("Fetch database with query <query>", async (query) => {
+const requestDatabase = async (query) => {
   return fetchFacade(`/api/database.json${query}`)
     .then(res => {
       assert(res.ok, `Could not fetch database`)
       return res.json()
     })
     .then(data => gauge.dataStore.scenarioStore.put('fetched.database', data))
+}
+
+step("Fetch database", async () => {
+  return requestDatabase('')
+})
+
+step("Fetch database with query <query>", async (query) => {
+  return requestDatabase(query)
 })
 
 step("Fetched database has <amount> entries", async (amount) => {