+++ packages/server/src/plugins.js
@@ -6,5 +6,6 @@ const __dirname = fileURLToPath(new URL('.', import.meta.url))
 export const getPluginFiles = () => {
   return [
     path.resolve(__dirname, 'api/database/queryTextCache.js'),
+    path.resolve(__dirname, 'auth/userFilter.js'),
   ]
 }